+++ DTCD-PrimitivePropertiesPanel/src/Plugin.js
@@ -16,16 +16,25 @@ export class Plugin extends PanelPlugin {
   constructor (guid, selector) {
     super();
 
-    const VueJS = this.getDependence('Vue');
     const logSystem = new LogSystemAdapter();
+    logSystem.info(`Start instantiation of ${pluginMeta.name} plugin`);
+    logSystem.debug(`Create LogSystemAdapter instance in ${pluginMeta.name} plugin`);
+
+    const VueJS = this.getDependence('Vue');
+    logSystem.debug(`Get Vue dependence in ${pluginMeta.name} plugin`);
+
     const eventSystem = new EventSystemAdapter();
+    logSystem.debug(`Create EventSystemAdapter instance in ${pluginMeta.name} plugin`);
 
     const data = { guid, logSystem, eventSystem };
 
+    logSystem.debug(`Creating Vue instance in ${pluginMeta.name} plugin`);
     new VueJS.default({
       data: () => data,
       render: h => h(PluginComponent),
     }).$mount(selector);
+
+    logSystem.info(`End of instantiation of ${pluginMeta.name} plugin`);
   }
 
 }